updated to d.js v14, finished reaction commands, and redid permissions

+++ commands/clear.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder } = require('@discordjs/builders');
+const { SlashCommandBuilder } = require('discord.js');
 const db = require('quick.db');
 
 module.exports = {
@@ -6,12 +6,18 @@ module.exports = {
 	.setName('clear')
 	.setDescription('Clears all current cooldowns, useful if the bot goes down'),
 	async execute(interaction){
-		let channel = interaction.channelId;
+		console.log(interaction.member._roles)
 
-		let state = db.get(`${channel}.paused`);
-		db.delete(channel);
-		db.set(`${channel}.paused`, state);
+		if(['767933568940638248', '880330557014294631', '461683459484549121', '875943961146032158'].some(v => interaction.member._roles.includes(v))){ // mod, supermod, dev, robotics
+			let channel = interaction.channelId;
 
-		await interaction.reply({ content: 'Cleared all level command cooldowns in all channels', ephemeral: true });
+			let state = db.get(`${channel}.paused`);
+			db.delete(channel);
+			db.set(`${channel}.paused`, state);
+
+			await interaction.reply({ content: 'Cleared all level command cooldowns in all channels', ephemeral: true });
+		}else{
+			return interaction.reply({ content: 'You do not have the adequate permissions to perform this action!', ephemeral: true });
+		}
 	}
 }
